Minor: comments, indentation, compilation warnings...
git-svn-id: https://svn.isais.edf.fr/mfee/saturne/trunk@10783 e44c5d27-f830-0410-8ce8-974f7a764be6

+++ src/cdo/cs_cdo_local.h
@@ -452,7 +452,7 @@ cs_cell_mesh_get_f2v(short int                    f,
   /* Tag vertices belonging to the current face f */
   for (short int i = cm->f2e_idx[f]; i < cm->f2e_idx[f+1]; i++) {
 
-    const short int  shift_e = 2*cm->f2e_ids[i];
+    const short int  shift_e = (short int)2*cm->f2e_ids[i];
     v_ids[cm->e2v_ids[shift_e]] = 1;
     v_ids[cm->e2v_ids[shift_e+1]] = 1;
